PASO 9: `EmployeeService` e `EmployeeServiceImpl`

+++ src/main/java/dev/magadiflo/app/repository/EmployeeRepository.java
@@ -4,7 +4,7 @@ import dev.magadiflo.app.entity.Employee;
 import io.reactivex.rxjava3.core.Observable;
 import org.springframework.data.repository.reactive.RxJava3CrudRepository;
 
-public interface EmployeeRepository extends RxJava3CrudRepository<Employee, Integer> {
+public interface EmployeeRepository extends RxJava3CrudRepository<Employee, Long> {
     Observable<Employee> findByPosition(String position);
 
     Observable<Employee> findByFullTime(Boolean isFullTime);
